+++ changelog.md
@@ -3,6 +3,7 @@ latest
  - Prev / Next buttons on the edit panel to navigate between samples without needing to close the edit panel and re-select from the list.
  - Cleaned up selected actions list, added overflow so list can scroll on mobile and flex to the width of the panel.
  - Shift + Ctrl/Cmd click stretch longest/shortest selected action to set the stretch length in samples instead of seconds.
+ - Shift + Click on Trim Right button in editor, or selected actions, to also trim silence from the start (left) of the sample.
 
 2024.02.20 1.4.11
  - Fixes bug where saving slice data from imported slice file did not apply slice markers consistently.